Fully qualify bundle install path, to fix issue when building on CodeBuild. Fixes #80.
When running jets deploy on CodeBuild, the bundled gems are not being copied to the correct location. That means that the zips that are produced and deployed contain no dependencies, resulting in lots of errors (e.g. "Could not find rake-12.3.1 in any of the sources"). This change fully qualifies the path, as the terminal on CodeBuild does not seem to be observing the cd command. Should be no functional change on existing build systems.
Testing:
- rspec
- Ran jets build and confirmd the zip contained bundled/gems as expected.

https://github.com/tongueroo/jets/issues/80