Fix route params for specs
Currenty, using named routes or having route parameters in specs did not work. For instance:

```ruby
get '/spec_helper_test/123' # route not found
get helper_path(123) # route not found

get '/spec_helper_test/:id', id: 123 # works
```

The latter syntax is quite awkward and not intuitive.

This commit extracts the route matcher and finder in dedicated classes
to reuse them in the middleware as well as in the spec helper.