Fix ApiResources with long names not in template
When ApiResource has a long name, it is being truncated by the
generator to match ApiGateway requirements of having resource names
no longer than 64 (62) characters.

Jets::Resource::ChildStack#controller_params when generating
templates is identifying resource ids by postfix, like Resource,
Authorizer or RootResourceId. Truncating was removing that postfix,
making it impossible to be properly recognized by that mechanism and
in effect, a failed deployment.

This fix changes the truncating code to make it shorter and keep
the postfix.